+++ ApiLogin/src/main/java/com/app/apilogin/integration/UsuarioWs.java
@@ -6,6 +6,6 @@ import com.app.apilogin.excepciones.WSException;
 import com.app.apilogin.modelo.Usuario;
 
 public interface UsuarioWs {
-	public Usuario obtenerUsuarioPorId(Long id) throws WSException;
-	public ResponseEntity<?> actualizarUsuarioPorID(Long id,Usuario usuario) throws WSException;
+	public Usuario obtenerUsuarioPorId(Integer id) throws WSException;
+	public ResponseEntity<?> actualizarUsuarioPorID(Integer id,Usuario usuario) throws WSException;
 }
