+++ ApiLogin/src/main/java/com/app/apilogin/repositorio/PagoRepositorio.java
@@ -5,10 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.app.apilogin.modelo.Pago;
 
 @Repository
-public interface PagoRepositorio extends JpaRepository<Pago, Long>{
-
-//    @Query(value = "SELECT * FROM Usuarios p WHERE p.email LIKE %:email% AND p.password LIKE %:pass%",
-//    nativeQuery = true)
-//    Usuario validarDatos(@Param(value = "email") String email, @Param(value = "pass") String pass);
+public interface PagoRepositorio extends JpaRepository<Pago, Integer>{
     
 }
