+++ ApiLogin/src/main/java/com/app/apilogin/service/PagoService.java
@@ -8,9 +8,6 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 @org.springframework.stereotype.Service
 public interface PagoService {
 
-//	public ResponseEntity<?> validarDatos(String email, String pass) throws JsonProcessingException;
-//	public ResponseEntity<?> actualizarUsuarioPorID(Long id,Usuario usuarioAct) throws JsonProcessingException;
-//	public ResponseEntity<?> guardarUsuario(Usuario usuario) throws JsonProcessingException;
-//	public ResponseEntity<?> obtenerUsuarioPorID(@PathVariable Long id) throws JsonProcessingException;
-	public ResponseEntity<?> realizarPago(Pago pago, Long id)throws JsonProcessingException;
+//	public ResponseEntity<?> obtenerPagoPorID(@PathVariable Long id) throws JsonProcessingException;
+	public ResponseEntity<?> realizarPago(Pago pago, Integer id)throws JsonProcessingException;
 }
